% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_comments_by_id.R
\name{get_comments_by_id}
\alias{get_comments_by_id}
\title{Get comments for one video based on its video ID.}
\usage{
get_comments_by_id(x)
}
\arguments{
\item{x}{A string}
}
\value{
All the comments on the video with video ID. of \code{x}
}
\description{
Get comments for one video based on its video ID.
}
\examples{
get_comments_by_id(SwBA1qb5ENg)
}
