% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_single_video_stats.R
\name{get_single_video_stats}
\alias{get_single_video_stats}
\title{Get Video Stats for one video based on its video ID.}
\usage{
get_single_video_stats(x = "SwBA1qb5ENg")
}
\arguments{
\item{x}{A string}
}
\value{
The video Stats for the video with video ID. of \code{x}
}
\description{
Get Video Stats for one video based on its video ID.
}
