\name{joined_datasets}
\alias{Joined_datasets}
\docType{data}
\title{
Joined Dateset 1 & 2 }
\description{
join two data sets by video ID and link video performance data and other data together.
}
\usage{data("joined_datasets")}
\format{
  A data frame with 226 observations on the following 14 variables.
}
