% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_word_cloud.R
\name{print_word_cloud}
\alias{print_word_cloud}
\title{Print word cloud for comments.}
\usage{
print_word_cloud(x = Comments_1$textOriginal, k = 30)
}
\arguments{
\item{x}{A text vector.}

\item{k}{A number.}
}
\value{
The wordcloud of \code{x} with words appears more than \code{k} times in \code{x}
}
\description{
Print word cloud for comments.
}
